requirements.txt
streamlit
modal
python-docx
